refactor(comments): code is now massively less hackish
At most a single optional comment controller is instantiated now, compared with
two from before.

The isolate scopes are now being made use of more correctly, rather than eval-
uating expressions from the parent scope. This simplifies the code somewhat, and
makes it considerably easier to grasp for beginners.

In general the library has been simplified somewhat, and should be that much more
effective from now on.

It is still desirable to make use of the sortBy attributes, so that more can be
done.