docs(demo): add Reddit browser to demo page
The CSS is pretty hack-ish at the moment, so bare with me. It simply shows off
the comment system using a real-world example. Due to not wanting to spam the
reddit API too much, lots of cacheing is happening, and stuff.

Unfortunately the code does depend on timeouts in some places, and is thus
potentially flaky :( There isn't much I can do about that just yet, but a fix
may be available soon due to similar issues with the $animate api in RC3.