refactor(comments): don't try to wait before $filled event
Previously, the app would use a $watcher, which would only be
cancelled when the child count of the compiled comments container
was greater than or equal to the parents length.

However, this wasn't really helpful, and was actively harmful in
cases where the repeated comments would not appear dirrectly
underneath the root of the template.

As such, this code has been drastically simplified, and the demo
improved to remove the pointless waiting, as well.